refactor: :hammer: refactor in halpers, infra and views
