+++ app/src/main/java/com/rodrigo/deeplarva/helpers/PreferencesHelper.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.utils
+package com.rodrigo.deeplarva.helpers
 
 import android.content.Context
 import android.content.SharedPreferences
