--- app/src/main/res/layout/dialog_add_picture.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:padding="16dp">
-
-    <TextView
-        android:id="@+id/tv_title_2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Sube un archivo o tome una foto con la cámara"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="16dp" />
-
-    <Button
-        android:id="@+id/btn_load_pic"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Subir archivo"
-        android:layout_below="@id/tv_title_2"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true" />
-
-    <Button
-        android:id="@+id/btn_camera"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Capturar"
-        android:layout_below="@id/btn_load_pic"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true" />
-
-</LinearLayout>
